+++ crates/lua/tests/scripts/input.lua
@@ -20,15 +20,12 @@ local kernel = utils.start_kernel("python3")
 local received_input_request = false
 local received_value = ""
 
-for res in kernel:execute("v = input('ASK> '); print('GOT:' + v)", 20) do
-	local msg = res.msg
-	if res.status == "busy" then
-		if msg.header.msg_type == "input_request" then
-			received_input_request = true
-			kernel:provide_stdin("", "bananas")
-		elseif msg.header.msg_type == "stream" and msg.content and msg.content.text then
-			received_value = received_value .. msg.content.text
-		end
+for msg in kernel:execute("v = input('ASK> '); print('GOT:' + v)", 20) do
+	if msg.header.msg_type == "input_request" then
+		received_input_request = true
+		kernel:provide_stdin("", "bananas")
+	elseif msg.header.msg_type == "stream" and msg.content and msg.content.text then
+		received_value = received_value .. msg.content.text
 	end
 end